sample/pyproject.toml
[tool.poetry]
name = "sample"
version = "0.1.0"
description = ""
authors = ["Brian McGrath <bmcgrath@grata.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
celery = "^5.5.3"
pydantic = "^2.11.7"
redis = "^5.0.0"
elasticsearch-dsl = "^8.0.0"
pytz = "^2025.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
